+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 17.1.1
+
+- Fixes formatting of error message when an invalid `types` value is specified.
+
+## 17.1.0
+
+- Allows for filtering by subtypes (e.g. `poi.landmark`) which are defined by score range.
+
 ## 17.0.0
 
 - Allow more flexible regexes in global tokens and refactor how they are applied.